files moved from PC to git,
A rest API i created ,its  using only the main package to reduce complexity and focus on CRUD(create, read, update, delete) operations.

main.go : This file is the starting point of the app. It contains all the code for DB connection and routing of API endpoints.

db-connection.go : This file contains the DB connection and migration code.

employee.go : this file contains the model definition or entity structure, which we want to store in our DB.

handlers.go : This contains all the logic for CRUD operations.

routing.go : this contains all the endpoints for the API within this app.